마지막 비밀번호 변경 날짜로부터 90일이 넘을 경우 비밀번호 변경 권유하기

+++ src/main/java/com/sbs/cyj/readit/controller/MemberController.java
@@ -40,6 +40,8 @@ public class MemberController {
 			if (id > 0) {
 				String code = memberService.genMailAuthCode(id);
 				memberService.sendJoinCompleteMail(param, code);
+				String now = memberService.getNowTime();
+				attrService.setValue("member__" + id + "__extra__lastPasswordChangeDate", now, "2030-01-01 01:01:01");
 			}
 			
 			return "<script> alert('" + id + "번째 회원입니다.'); location.replace('../home/main'); </script>";
@@ -77,24 +79,18 @@ public class MemberController {
 			redirectUri = "/usr/home/main";
 		}
 		
-		String str = "안녕하세요.";
+		String str = String.format("%s님, 안녕하세요.", member.getNickname());
 
-		//		String lastUpdateDate = member.getUpdateDate();
-//		
-//		int date = memberService.compareDate(lastUpdateDate);
-//		if(date > 90) {
-//			str = String.format("%s님 비밀번호 변경한 지 %d일이 되었습니다.\\n보완을 위해 비밀번호를 바꿔주세요.", member.getNickname(), date);
-//		}
-//		else {
-//			str = String.format("%s님 반갑습니다.", member.getNickname());
-//		}
-//		
-//		if(member.isTempPwStatus()) {
-//			str += "\\n임시번호를 사용중입니다.\\n비밀번호 재설정을 해주세요.";
-//		}
-		
 		if(attrService.getValue("member__" + member.getId() + "__extra__useTempPassword").equals("1")) {
 			str = "현재 임시 비밀번호를 사용중입니다.";
+		}
+		
+		if(attrService.getValue("member__" + member.getId() + "__extra__lastPasswordChangeDate").length() != 0) {
+			String lastUpdateDate = attrService.getValue("member__" + member.getId() + "__extra__lastPasswordChangeDate");
+			int date = memberService.compareDate(lastUpdateDate); 
+			if( date > 90) {
+				str = String.format("%s님 비밀번호 변경한 지 %d일이 되었습니다.\\n보완을 위해 비밀번호를 바꿔주세요.", member.getNickname(), date);
+			}
 		}
 
 		model.addAttribute("redirectUri", redirectUri);
@@ -313,7 +309,7 @@ public class MemberController {
 		return "common/redirect";
 	}
 	
-	// 회원 정보 
+	// 회원 정보 수정
 	@RequestMapping("/usr/member/modify")
 	public String showModify(HttpSession session, HttpServletRequest req, String checkPasswordAuthCode, Model model) {
 		int memberId = (int) session.getAttribute("loginedMemberId");
@@ -344,8 +340,15 @@ public class MemberController {
 			param.put("email", null);
 		}
 		
-		if(attrService.getValue("member__" + member.getId() + "__extra__useTempPassword").equals("1")) {
-			attrService.remove("member__" + memberId + "__extra__useTempPassword");
+		String newPw = (String) param.get("newLoginPwReal"); 
+		
+		if(newPw.length() != 0) {
+			if(attrService.getValue("member__" + member.getId() + "__extra__useTempPassword").equals("1")) {
+				attrService.remove("member__" + memberId + "__extra__useTempPassword");
+			}
+			
+			String now = memberService.getNowTime();
+			attrService.setValue("member__" + member.getId() + "__extra__lastPasswordChangeDate", now, "2030-01-01 01:01:01");
 		}
 		
 		memberService.modify(param);
